Migrating library Export to RichSkillController, removing OsmtUtils class, refactoring to not use StreamingResponseBody
